+++ test/unit/test_trainer.py
@@ -13,7 +13,7 @@
 import errno
 import os
 
-from mock import Mock, patch
+from mock import MagicMock, Mock, patch
 
 from sagemaker_containers import _errors, _trainer
 
@@ -22,11 +22,26 @@ class TrainingEnv(Mock):
     framework_module = 'my_framework:entry_point'
     log_level = 20
 
+    def sagemaker_s3_output(self):
+        return 's3://bucket'
 
-class SriptTrainingEnv(TrainingEnv):
+
+class TrainingEnvNoIntermediate(Mock):
+    framework_module = 'my_framework:entry_point'
+    log_level = 20
+
+    def sagemaker_s3_output(self):
+        return None
+
+
+class ScriptTrainingEnv(TrainingEnv):
     framework_module = None
 
+    def sagemaker_s3_output(self):
+        return 's3://bucket'
 
+
+@patch('inotify_simple.INotify', MagicMock())
 @patch('importlib.import_module')
 @patch('sagemaker_containers.training_env', TrainingEnv)
 def test_train(import_module):
@@ -38,6 +53,7 @@ def test_train(import_module):
     framework.entry_point.assert_called()
 
 
+@patch('inotify_simple.INotify', MagicMock())
 @patch('importlib.import_module')
 @patch('sagemaker_containers.training_env', TrainingEnv)
 @patch('sagemaker_containers._trainer._exit_processes')
@@ -53,6 +69,7 @@ def test_train_with_success(_exit, import_module):
     _exit.assert_called_with(_trainer.SUCCESS_CODE)
 
 
+@patch('inotify_simple.INotify', MagicMock())
 @patch('importlib.import_module')
 @patch('sagemaker_containers.training_env', TrainingEnv)
 @patch('sagemaker_containers._trainer._exit_processes')
@@ -69,6 +86,7 @@ def test_train_fails(_exit, import_module):
     _exit.assert_called_with(errno.ENOENT)
 
 
+@patch('inotify_simple.INotify', MagicMock())
 @patch('importlib.import_module')
 @patch('sagemaker_containers.training_env', TrainingEnv)
 @patch('sagemaker_containers._trainer._exit_processes')
@@ -85,8 +103,9 @@ def test_train_with_client_error(_exit, import_module):
     _exit.assert_called_with(_trainer.DEFAULT_FAILURE_CODE)
 
 
+@patch('inotify_simple.INotify', MagicMock())
 @patch('sagemaker_containers.entry_point.run')
-@patch('sagemaker_containers.training_env', new_callable=SriptTrainingEnv)
+@patch('sagemaker_containers.training_env', new_callable=ScriptTrainingEnv)
 @patch('sagemaker_containers._trainer._exit_processes')
 def test_train_script(_exit, training_env, run):
     _trainer.train()
@@ -96,3 +115,16 @@ def test_train_script(_exit, training_env, run):
                            env.to_env_vars())
 
     _exit.assert_called_with(_trainer.SUCCESS_CODE)
+
+
+@patch('importlib.import_module')
+@patch('sagemaker_containers._intermediate_output.start_sync')
+@patch('sagemaker_containers.training_env', TrainingEnvNoIntermediate)
+def test_train_no_intermediate(start_intermediate_folder_sync, import_module):
+    framework = Mock()
+    import_module.return_value = framework
+    _trainer.train()
+
+    import_module.assert_called_with('my_framework')
+    framework.entry_point.assert_called()
+    start_intermediate_folder_sync.asser_not_called()